refactor: 메서드 변경
