+++ backend/src/main/java/hanglog/expense/service/ExpenseService.java
@@ -6,11 +6,11 @@ import static hanglog.global.exception.ExceptionCode.NOT_FOUND_TRIP_ID;
 import hanglog.category.domain.Category;
 import hanglog.category.domain.repository.CategoryRepository;
 import hanglog.currency.domain.Currency;
+import hanglog.currency.domain.repository.CurrencyRepository;
 import hanglog.currency.domain.type.CurrencyType;
 import hanglog.expense.domain.CategoryExpense;
 import hanglog.expense.domain.DayLogExpense;
 import hanglog.expense.domain.Expense;
-import hanglog.expense.domain.repository.CurrencyRepository;
 import hanglog.expense.dto.response.TripExpenseResponse;
 import hanglog.global.exception.BadRequestException;
 import hanglog.trip.domain.DayLog;
