+++ backend/src/test/java/hanglog/currency/presentation/CurrencyControllerTest.java
@@ -1,13 +1,9 @@
 package hanglog.currency.presentation;
 
-import static hanglog.expense.fixture.CurrencyFixture.DEFAULT_CURRENCY;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.when;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 import hanglog.currency.service.CurrencyService;
-import java.time.LocalDate;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -29,10 +25,6 @@ class CurrencyControllerTest {
     @DisplayName("당일 환율 정보를 저장한다.")
     @Test
     void saveTodayCurrency() throws Exception {
-        // given
-        when(currencyService.saveDailyCurrency(any(LocalDate.class)))
-                .thenReturn(DEFAULT_CURRENCY);
-
         // when & then
         mockMvc.perform(post("/currency/today"))
                 .andExpect(status().isOk());
